requirements.txt
anthropic>=0.40.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
anyio>=4.4.0
loguru>=0.7.2
rich>=13.7.0
typer>=0.12.0
aiosqlite>=0.20.0
httpx>=0.27.0
pymetasploit3>=1.0.3
dnspython>=2.6.0
python-dotenv>=1.0.0
jinja2>=3.1.0
weasyprint>=62.0
